fix: simplify task completion CLI output
